LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/contain-paint-offscreen-container-expected.txt


FAIL Contain: style paint container offscreen. assert_equals: expected -2 but got -102
